more type changes, some consts added
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@1061 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/BufferView.h
@@ -192,11 +192,11 @@ public:
 	///
 	bool lockInset(UpdatableInset * inset);
 	///
-	void showLockedInsetCursor(long x, long y, int asc, int desc);
+	void showLockedInsetCursor(int x, int y, int asc, int desc);
 	///
 	void hideLockedInsetCursor();
 	///
-	void fitLockedInsetCursor(long x, long y, int asc, int desc);
+	void fitLockedInsetCursor(int x, int y, int asc, int desc);
 	///
 	int unlockInset(UpdatableInset * inset);
 	///
